Track both line and column in the tokenizer.
This involves a rather large refactor that I'm very pleased with. We
remove the Eof token entirely and instead rely on Rust's optional type
returning None to indicate that the tokenizer is out of tokens. The
advantage of this is that because we're leaning on the type system,
every handler of tokens must deal with the possibility that there are
no more. I beleive that this will reduce the number of Eof related
errors we see down the line.

This commit also introduces a Reader struct that handles getting
characters out of a string and tracking our line and column
number. This is nice because it allows us to abstract location
tracking away entirely and handle weird cases like tabs and newlines
as we go. The reader currently calculates the visual column number as
well so that errrors can be sure to underline the correct seciton of
text even if a literal tab character is encountered.