Add some basic numerical operators to the interpreter.
At this point I beleive that I have, once again, hit a roadblock for
this implementation. The way that I'm handling everything being a
anonymous function call means that I'm having trouble working out how
to do tail recursion. It essentially goes like this:

With if being an builtin anonymous function it is called just like
anything else and then depending on the value of its condition
evaluates its arguments. We'd like to make things tail recursive which
would involve instead of evaluating the arguments by calling back into
the interpreter, we'd evaluate them by going back to the top of the
evaluation loop. The trouble is that we don't know where the
evaulation loop is if we're in the if function call.

We can detect certian kinds of recursion though by looking to see if
the last item in a function's body is a function call. Sadly this
isn't terribly interesting because we pretty much need conditionals to
be tail recrusive to do most useful things.