Strings are just lists of characters
Previously strings were c strings. This was nice for making foreign
calls but makes strings a whole new language feature that needs to be
learned. Instead we now treat strings as regular lists of characters.

This commit also adds a printf function which works the way you'd
expect it to using format strings from rust. For example
`(printf "hello {}" "zeke")` prints "hello zeke".

This is a very exciting milestone because a printf function was the
last thing we wrote in the interpreter before moving on to the
compiler. Outside of memory management the compiler is now very close
to feature parity with the interpreter!